fix(tests): fix tests for alpha
