+++ apps/alpha/src/db/tenancy/tenant-db-connection.ts
@@ -1,39 +1,9 @@
 import { BadRequestError } from 'fastify-custom-errors';
-import type { Options } from 'sequelize';
 import { QueryTypes } from 'sequelize';
 import type { CacheTenantRecord, TenantRecords } from '../../cache';
 import { MainCacheInstance } from '../../cache';
 import { SequelizeInstance } from '../instance';
-
-/**
- * Connection to the main database which stores the tenants
- * and tenant related metadata.
- *
- * @param [opts] - Options bag to forward to sequelize
- * @returns Sequelize connection
- */
-export function getMainDBConnection(opts?: Options) {
-  if (!process.env.DB_DATABASE) {
-    throw new Error('Expected `DB_DATABASE` to be defined but was not set!');
-  }
-
-  if (!process.env.DB_USERNAME) {
-    throw new Error('Expected `DB_USERNAME` to be defined but was not set!');
-  }
-
-  if (!process.env.DB_PASSWORD) {
-    throw new Error('Expected `DB_PASSWORD` to be defined but was not set!');
-  }
-
-  const _conn = new SequelizeInstance(
-    process.env.DB_DATABASE,
-    process.env.DB_USERNAME,
-    process.env.DB_PASSWORD,
-    opts,
-  );
-
-  return _conn;
-}
+import { getMainDBConnection } from './main-db-connection';
 
 /**
  * Connect to the tenant db with the provided tenant creds
